[3scale] rename APIcast policy to 3scale
