+++ examples/configuration/echo.json
@@ -15,7 +15,7 @@
           "host": "echo"
         },
         "policy_chain": [
-          { "name": "apicast.policy.apicast" }
+          { "name": "apicast.policy.3scale" }
         ],
         "proxy_rules": [
           {
