+++ gateway/src/apicast/policy/3scale/apicast-policy.json
