+++ gateway/src/apicast/policy/3scale/init.lua
@@ -1,0 +1,1 @@
+return require('3scale')
