+++ gateway/src/apicast/policy/apicast/init.lua
@@ -1,1 +1,3 @@
-return require('apicast')
+ngx.log(ngx.WARN, 'DEPRECATION: file renamed - change: require("apicast.policy.apicast")' ,' to: require("apicast.policy.3scale")')
+
+return require('apicast.policy.3scale')
