+++ spec/examples/custom-module/blacklist_spec.lua
@@ -1,5 +1,5 @@
 local _M = require 'examples.custom-module.blacklist'
-local apicast = require 'apicast.policy.apicast'
+local apicast = require 'apicast.policy.3scale'
 
 describe('blacklist', function()
   it('returns new module instance', function()
